Data Binding for General Tab (#VR-68, #VR66, #VR67, #VR71, #VR72)
Most of the UI-Elements on the General Tab are now filled with the correct data!

+++ QtQuickApplication/view/vrplotpedals.cpp
@@ -5,7 +5,7 @@ VRPlotPedals::VRPlotPedals() : VRPlotItem()
 
 }
 
-void VRPlotPedals::pushData(int clutch, int brake, int accel)
+void VRPlotPedals::pushData(double clutch, double brake, double accel)
 {
     QVector<double> ticks_vect_clutch;
     QVector<double> ticks_vect_brak;
